# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ParkWell - Parking Attendant System API",
    "description": "API documentation for the ParkWell system, for administrators and parking attendants.",
    "version": "1.1.0",
    "contact": {
      "name": "ParkWell Support",
      "email": "support@parkwell.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development Server (v1)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/reports/exited-vehicles": {
      "get": {
        "tags": [
          "Admin - Reports"
        ],
        "summary": "Get exited vehicles report",
        "operationId": "getExitedVehiclesReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageQueryParam"
          },
          {
            "$ref": "#/components/parameters/LimitQueryParamReports"
          },
          {
            "$ref": "#/components/parameters/SortByExitedReportQueryParam"
          },
          {
            "$ref": "#/components/parameters/OrderQueryParam"
          },
          {
            "$ref": "#/components/parameters/StartDateQueryParam"
          },
          {
            "$ref": "#/components/parameters/EndDateQueryParam"
          },
          {
            "$ref": "#/components/parameters/ParkingIdQueryParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Exited vehicles report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExitedVehiclesReportResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/admin/reports/entered-vehicles": {
      "get": {
        "tags": [
          "Admin - Reports"
        ],
        "summary": "Get entered vehicles report",
        "operationId": "getEnteredVehiclesReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageQueryParam"
          },
          {
            "$ref": "#/components/parameters/LimitQueryParamReports"
          },
          {
            "$ref": "#/components/parameters/SortByEnteredReportQueryParam"
          },
          {
            "$ref": "#/components/parameters/OrderQueryParam"
          },
          {
            "$ref": "#/components/parameters/StartDateQueryParam"
          },
          {
            "$ref": "#/components/parameters/EndDateQueryParam"
          },
          {
            "$ref": "#/components/parameters/ParkingIdQueryParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Entered vehicles report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnteredVehiclesReportResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Specific error message"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "example": 10
          },
          "totalItems": {
            "type": "integer",
            "example": 95
          },
          "itemsPerPage": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
      "VehicleEntryStatusEnum": {
        "type": "string",
        "enum": [
          "PARKED",
          "EXITED"
        ],
        "example": "PARKED"
      },
...
      "VehicleEntryDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plate_number": {
            "type": "string"
          },
          "entry_time": {
            "type": "string",
            "format": "date-time"
          },
          "exit_time": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "calculated_duration_minutes": {
            "type": "integer",
            "nullable": true
          },
          "charged_amount": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/VehicleEntryStatusEnum"
          },
          "ticket_number": {
            "type": "string"
          },
          "parking_id": {
            "type": "string",
            "format": "uuid"
          },
          "parking": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "charge_per_hour": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "recorded_by_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "recorded_by": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReportSummaryExited": {
        "type": "object",
        "properties": {
          "totalVehiclesExited": {
            "type": "integer"
          },
          "totalRevenue": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "ReportSummaryEntered": {
        "type": "object",
        "properties": {
          "totalVehiclesEntered": {
            "type": "integer"
          }
        }
      },
      "ExitedVehiclesReportResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleEntryDetails"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/ReportSummaryExited"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "EnteredVehiclesReportResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleEntryDetails"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/ReportSummaryEntered"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      }
    },
    "parameters": {
...
      "PageQueryParam": {
        "name": "page",
        "in": "query",
        "description": "Page number",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
...
      "LimitQueryParamReports": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page for reports",
        "schema": {
          "type": "integer",
          "default": 15,
          "minimum": 1,
          "maximum": 100
        }
      },
      "OrderQueryParam": {
        "name": "order",
        "in": "query",
        "description": "Sort order (asc/desc)",
        "schema": {
          "type": "string",
          "enum": [
            "asc",
            "desc"
          ],
          "default": "desc"
        }
      },
...
      "SortByExitedReportQueryParam": {
        "name": "sortBy",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "exit_time",
            "plate_number",
            "charged_amount",
            "calculated_duration_minutes",
            "parking.name"
          ],
          "default": "exit_time"
        }
      },
      "SortByEnteredReportQueryParam": {
        "name": "sortBy",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "entry_time",
            "plate_number",
            "status",
            "parking.name"
          ],
          "default": "entry_time"
        }
      },
      "StartDateQueryParam": {
        "name": "startDate",
        "in": "query",
        "description": "Format YYYY-MM-DD",
        "schema": {
          "type": "string",
          "format": "date"
        }
      },
      "EndDateQueryParam": {
        "name": "endDate",
        "in": "query",
        "description": "Format YYYY-MM-DD",
        "schema": {
          "type": "string",
          "format": "date"
        }
      },
      "ParkingIdQueryParam": {
        "name": "parkingId",
        "in": "query",
        "description": "Filter by Parking Facility UUID",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized. Token missing or invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden. Insufficient permissions.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "BadRequestError": {
        "description": "Bad Request. Invalid input or validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}